+++ twitter4j-v2-support/src/main/kotlin/twitter4j/ListsEx.kt
@@ -67,3 +67,70 @@ fun Twitter.deleteList(
         "deleted"
     )
 }
+
+/**
+ * Enables the authenticated user to add a member to a List they own.
+ *
+ * @throws TwitterException when Twitter service or network is unavailable
+ * @see "https://developer.twitter.com/en/docs/twitter-api/lists/manage-lists/api-reference/post-lists-id-members"
+ */
+@Throws(TwitterException::class)
+fun Twitter.addListMember(
+    /**
+     * The ID of the List you are adding a member to.
+     */
+    listId: Long,
+    /**
+     * The ID of the user you wish to add as a member of the List.
+     */
+    userId: Long
+): BooleanResponse {
+
+    if (this !is TwitterImpl) throw IllegalStateException("invalid twitter4j impl")
+
+    ensureAuthorizationEnabled()
+
+    val json = JSONObject()
+    json.put("user_id", userId.toString())
+
+    return V2ResponseFactory().createBooleanResponse(
+        http.post(
+            conf.v2Configuration.baseURL + "lists/" + listId + "/members",
+            arrayOf(HttpParameter(json)),
+            auth,
+            this
+        ),
+        conf,
+        "is_member"
+    )
+}
+
+/**
+ * Enables the authenticated user to remove a member from a List they own.
+ *
+ * @throws TwitterException when Twitter service or network is unavailable
+ * @see "https://developer.twitter.com/en/docs/twitter-api/lists/manage-lists/api-reference/delete-lists-id-members-user_id"
+ */
+@Throws(TwitterException::class)
+fun Twitter.deleteListMember(
+    /**
+     * The ID of the List you are removing a member from.
+     */
+    listId: Long,
+    /**
+     * The ID of the user you wish to remove as a member of the List.
+     */
+    userId: Long
+): BooleanResponse {
+
+    if (this !is TwitterImpl) throw IllegalStateException("invalid twitter4j impl")
+
+    ensureAuthorizationEnabled()
+
+    return V2ResponseFactory().createBooleanResponse(
+        http.delete(conf.v2Configuration.baseURL + "lists/" + listId + "/members/" + userId, emptyArray(), auth, this),
+        conf,
+        "is_member"
+    )
+}
+