+++ CHANGELOG.md
@@ -5,6 +5,8 @@ v0.3.2 (2021.10.xx)
 - Manage Lists
   - Add Twitter.createList() for "POST /2/lists"
   - Add Twitter.deleteList() for "DELETE /2/lists/:id"
+  - Add Twitter.addListMember() for "POST /2/lists/:id/members"
+  - Add Twitter.deleteListMember() for "DELETE /2/lists/:id/members/:user_id"
 
 v0.3.1 (2021.09.24)
 -------------------